feat: adicionando validações para PessoaVo, foi alterado o formato int para Long no id

+++ src/main/java/com/projetos/labH2/labDAO/PessoaDao.java
@@ -12,9 +12,9 @@ public interface PessoaDao {
 
     List<PessoaVo> getAllPessoas();
 
-    PessoaVo getPessoaById(@Param("id") int id);
+    PessoaVo getPessoaById(@Param("id") Long id);
 
-    void deletePessoaById(@Param("id") int id);
+    void deletePessoaById(@Param("id") Long id);
 
     PessoaVo updatePessoaById(PessoaVo pessoa);
 }
